+++ api/e_host/create.php
@@ -1,0 +1,68 @@
+<?php
+// required headers
+header("Access-Control-Allow-Origin: *");
+header("Content-Type: application/json; charset=UTF-8");
+header("Access-Control-Allow-Methods: POST");
+header("Access-Control-Max-Age: 3600");
+header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
+  
+// get database connection
+include_once '../config/database.php';
+  
+// instantiate product object
+include_once '../objects/e_hosts.php';
+  
+$database = new Database();
+$db = $database->getConnection();
+  
+$e_host = new E_host($db);
+  
+// get posted data
+$data = json_decode(file_get_contents("php://input"));
+  
+// make sure data is not empty
+if(
+    !empty($data->Name) &&
+    !empty($data->Date) 
+){
+  
+    // set product property values
+    $e_host->Name = $data->Name;
+    $e_host->Date = $data->Date;
+    $e_host->UCID = $data->UCID;
+
+
+
+
+  
+    // create the product
+    if($e_host->create()){
+  
+        // set response code - 201 created
+        http_response_code(201);
+  
+        // tell the user
+        echo json_encode(array("message" => "E_host was created."));
+    }
+  
+    // if unable to create the product, tell the user
+    else{
+  
+        // set response code - 503 service unavailable
+        http_response_code(503);
+  
+        // tell the user
+        echo json_encode(array("message" => "Unable to create e_host. Choose a unique e_host for the event or check that event is entered correctly."));
+    }
+}
+  
+// tell the user data is incomplete
+else{
+  
+    // set response code - 400 bad request
+    http_response_code(400);
+  
+    // tell the user
+    echo json_encode(array("message" => "Unable to create e_host. Data is incomplete."));
+}
+?>