+++ api/e_host/delete.php
@@ -1,0 +1,49 @@
+<?php
+    // required headers
+    header("Access-Control-Allow-Origin: *");
+    header("Content-Type: application/json; charset=UTF-8");
+    header("Access-Control-Allow-Methods: POST");
+    header("Access-Control-Max-Age: 3600");
+    header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
+  
+    // include database and object file
+    include_once '../config/database.php';
+    include_once '../objects/e_hosts.php';
+  
+    // get database connection
+    $database = new Database();
+    $db = $database->getConnection();
+  
+    // prepare product object
+    $e_host = new E_host($db);
+  
+    // get product id
+    $data = json_decode(file_get_contents("php://input"));
+  
+    // set product id to be deleted
+    $e_host->Name = $data->Name;
+    $e_host->Date = $data->Date;
+    $e_host->UCID = $data->UCID;
+
+
+  
+    // delete the product
+    if($e_host->delete()){
+  
+        // set response code - 200 ok
+        http_response_code(200);
+  
+        // tell the user
+        echo json_encode(array("message" => "E_host was deleted."));
+    }
+  
+    // if unable to delete the product
+    else{
+  
+        // set response code - 503 service unavailable
+        http_response_code(503);
+  
+        // tell the user
+        echo json_encode(array("message" => "Unable to delete e_host."));
+    }
+?>